refactor: 이미지 삽입 기능 추가 및 파일명 변경
